[refactor] App.js 다이어트
1. findDocumentDataById 분리
