+++ src/page/HomePage.js
@@ -4,6 +4,7 @@ import { request } from '../services/api.js';
 import { initRouter } from '../utils/router.js';
 import styleInJS from '../style/tagStyles.js';
 import { addSaveDocumentTitleEvent } from '../utils/saveDocumentTitle.js';
+import { findDocumentDataById } from '../utils/findDocumentsPathData.js';
 
 /*
  * HomePage
@@ -17,35 +18,14 @@ export default function HomePage({ $target }) {
   $target.appendChild($homePage);
 
   this.state = [];
+  const initDocumentData = { id: null, title: '첫 화면', content: '내용을 채워주세요', documentPath: [] };
+  let tempDocumentData;
 
   const notionSideBar = new NotionSideBar({ $target: $homePage, initialState: [] });
   const documentDetail = new DocumentDetail({
     $target: $homePage,
-    documentState: { id: null, title: '첫 화면', content: '내용을 채워주세요', documentPath: [] },
+    documentState: initDocumentData,
   });
-
-  const findDocumentDataById = id => {
-    const documentPath = [];
-
-    const findDocument = (documents, id) => {
-      for (const document of documents) {
-        if (document.id === id) {
-          documentPath.push({ title: document.title, id: document.id });
-          return true;
-        }
-        if (document.documents.length > 0) {
-          documentPath.push({ title: document.title, id: document.id });
-          if (findDocument(document.documents, id)) return true;
-          documentPath.pop();
-        }
-      }
-    };
-    findDocument(this.state, id);
-
-    return documentPath;
-  };
-
-  let tempDocumentData;
 
   this.handleRoute = async () => {
     const { pathname } = window.location;
@@ -56,7 +36,7 @@ export default function HomePage({ $target }) {
     if (pathname === '/') {
       // home 보여주기
       notionSideBar.setState(this.state);
-      documentDetail.setState({ id: null, title: '첫 화면', content: '내용을 채워주세요', documentPath: [] });
+      documentDetail.setState(initDocumentData);
     } else if (pathname.indexOf('/documents/') === 0) {
       const [, , postId] = pathname.split('/');
       const documentContent = request(`/documents/${postId}`);
